+++ serving/torchserve/llama3-8b/material/compile.py
@@ -1,0 +1,23 @@
+import os
+
+from optimum.rbln import RBLNLlamaForCausalLM
+
+
+def main():
+    model_id = "meta-llama/Meta-Llama-3-8B-Instruct"
+
+    # Compile and export
+    model = RBLNLlamaForCausalLM.from_pretrained(
+        model_id=model_id,
+        export=True,  # export a PyTorch model to RBLN model with optimum
+        rbln_batch_size=4,
+        rbln_max_seq_len=8192,  # default "max_position_embeddings"
+        rbln_tensor_parallel_size=4,
+    )
+
+    # Save compiled results to disk
+    model.save_pretrained(os.path.basename(model_id))
+
+
+if __name__ == "__main__":
+    main()