+++ serving/triton_inference_server/llama3.1-8b/setup.sh
@@ -13,7 +13,7 @@ echo "=================================================================="
 DIR="vllm_backend"
 GIT_REPO="https://github.com/triton-inference-server/${DIR}.git"
 if [ ! -d ${DIR} ]; then
-  git clone ${GIT_REPO} -b r24.12
+  git clone ${GIT_REPO} -b r25.08
   if [ $? -ne 0 ]; then
     echo "Error while cloning git repository."
     popd
@@ -29,7 +29,7 @@ echo "=================================================================="
 
 DIR="Llama-3.1-8B-Instruct"
 if [ ! -d ${DIR} ]; then
-  python3 ../material/get_model.py
+  python3 ../material/compile.py
   if [ $? -ne 0 ];then
     echo "An error occurred while compiling the model."
     popd